[Outline] make a workaround to resolve JQuery UI sortable issue
It is difficult to locate Widget in the right place using original
implementation of JQuery UI sortable. We made a workaround to make
JQuery UI sortable() function. This workaround try to decrease the
times of rearange() when widgets is sorting. This make user can
insert widget easily relate to original method.